devex - remove duplicate files and move to shared directory
