+++ cypress-smoketests.config.ts
@@ -3,7 +3,7 @@ import {
   getUserTokenWithRetry,
 } from './cypress/support/cognito-login';
 import { defineConfig } from 'cypress';
-import { getUserToken as getUserTokenLocal } from './cypress/cypress-smoketests/support/local-login';
+import { getUserToken as getUserTokenLocal } from './cypress/helpers/auth/local-login';
 import { waitForNoce } from './cypress/cypress-smoketests/support/wait-for-noce';
 const { CYPRESS_SMOKETESTS_LOCAL } = process.env;
