+++ cypress/cypress-readonly/integration/public/advanced-search.cy.ts
@@ -1,20 +1,12 @@
-import {
-  docketRecordTable,
-  enterPetitionerName,
-  navigateTo as navigateToDashboard,
-  noSearchResultsContainer,
-  searchForCaseByDocketNumber,
-} from '../../support/pages/public/advanced-search';
-
 import { isValidRequest } from '../../support/helpers';
 
 const EFCMS_DOMAIN = Cypress.env('EFCMS_DOMAIN');
 const DEPLOYING_COLOR = Cypress.env('DEPLOYING_COLOR');
 
-describe('Case Search Public UI Smoketests', () => {
+describe('advanced search pages', () => {
   it('should allow the user to search for a case by petitioner name', () => {
-    navigateToDashboard();
-    enterPetitionerName('Smith');
+    cy.visit('/');
+    cy.get('input#petitioner-name').type('Smith');
 
     cy.intercept({
       hostname: `public-api-${DEPLOYING_COLOR}.${EFCMS_DOMAIN}`,
@@ -28,20 +20,20 @@ describe('Case Search Public UI Smoketests', () => {
   });
 
   it('should display "No Matches Found" when case search yields no results', () => {
-    navigateToDashboard();
-    searchForCaseByDocketNumber('99-21');
-    noSearchResultsContainer().should('exist');
+    cy.visit('/');
+    cy.get('input#docket-number').type('99-21');
+    cy.get('button#docket-search-button').click();
+
+    cy.get('div#no-search-results').should('exist');
   });
 
   it('should route to case detail when a case search by docket number match is found', () => {
-    navigateToDashboard();
-    searchForCaseByDocketNumber('104-20');
-    const docketRecord = docketRecordTable();
-    docketRecord.should('exist');
-    const petitionRow = docketRecord
-      .get('tr')
-      .contains('Petition', { matchCase: false });
-    petitionRow.should('exist');
-    petitionRow.find('button').should('not.exist');
+    cy.visit('/');
+    cy.get('input#docket-number').type('104-20');
+    cy.get('button#docket-search-button').click();
+    cy.get('table#docket-record-table tr:contains("Petition")').should('exist');
+    cy.get('table#docket-record-table tr:contains("Petition") button').should(
+      'not.exist',
+    );
   });
 });
