+++ cypress/cypress-readonly/integration/public/todays-orders.cy.ts
@@ -3,18 +3,16 @@ import { isValidRequest } from '../../support/helpers';
 const EFCMS_DOMAIN = Cypress.env('EFCMS_DOMAIN');
 const DEPLOYING_COLOR = Cypress.env('DEPLOYING_COLOR');
 
-describe('Public UI Smoketests', () => {
-  describe('todays orders', () => {
-    it('should fetch today orders from the public api', () => {
-      cy.intercept({
-        hostname: `public-api-${DEPLOYING_COLOR}.${EFCMS_DOMAIN}`,
-        method: 'GET',
-        url: '/public-api/todays-orders/**',
-      }).as('getTodaysOrders');
+describe('todays orders', () => {
+  it('should fetch today orders from the public api', () => {
+    cy.intercept({
+      hostname: `public-api-${DEPLOYING_COLOR}.${EFCMS_DOMAIN}`,
+      method: 'GET',
+      url: '/public-api/todays-orders/**',
+    }).as('getTodaysOrders');
 
-      cy.visit('/todays-orders');
+    cy.visit('/todays-orders');
 
-      cy.wait('@getTodaysOrders').then(isValidRequest);
-    });
+    cy.wait('@getTodaysOrders').then(isValidRequest);
   });
 });
