+++ cypress/cypress-readonly/support/index.ts
@@ -1,4 +1,1 @@
-before(() => {
-  // Skip subsequent tests in spec when one fails.
-  (cy.state('runnable').ctx as Mocha.Context).currentTest.parent.bail(true);
-});
+import '../../support/commands';
