+++ cypress/helpers/auth/local-login.ts
@@ -1,4 +1,4 @@
-export const getUserToken = async (username: string) => {
+export const getUserToken = (username: string) => {
   if (username === 'migrator@example.com') {
     username =
       'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFkbWluIiwibmFtZSI6IlRlc3QgQWRtaW4iLCJyb2xlIjoiYWRtaW4iLCJ1c2VySWQiOiI4NmMzZjg3Yi0zNTBiLTQ3N2QtOTJjMy00M2JkMDk1Y2IwMDYiLCJjdXN0b206cm9sZSI6ImFkbWluIiwic3ViIjoiODZjM2Y4N2ItMzUwYi00NzdkLTkyYzMtNDNiZDA5NWNiMDA2IiwiaWF0IjoxNTgyOTIxMTI1fQ.PBmSyb6_E_53FNG0GiEpAFqTNmooSh4rI0ApUQt3UH8';
@@ -14,6 +14,6 @@ export const login = (email: string, route = '/') => {
   cy.visit(`/log-in?code=${email}&path=${route}`);
 };
 
-export const getRestApi = async () => {
+export const getRestApi = () => {
   return 'localhost:4000';
 };
