+++ cypress/helpers/auth/login-as-helpers.ts
@@ -1,0 +1,4 @@
+export function loginAsTestAdmissionsClerk() {
+  cy.login('testAdmissionsClerk');
+  cy.get('#inbox-tab-content').should('exist');
+}
