+++ cypress/support/commands.ts
@@ -1,5 +1,5 @@
 import 'cypress-file-upload';
-import { getEnvironmentSpecificFunctions } from './environment-specific-factory';
+import { getEnvironmentSpecificFunctions } from '../helpers/auth/environment-specific-factory';
 
 const { login } = getEnvironmentSpecificFunctions();
 
@@ -8,4 +8,9 @@ Cypress.Commands.add('login', (username, route = '/') => {
   cy.window().then(win =>
     win.localStorage.setItem('__cypressOrderInSameTab', 'true'),
   );
+});
+
+before(() => {
+  // Skip subsequent tests in spec when one fails.
+  (cy.state('runnable').ctx as Mocha.Context).currentTest?.parent?.bail(true);
 });
